cleaned up example api task
